style: auto-fix formatting (gofumpt, goimports, gofmt)
Run golangci-lint --fix to apply automated formatting fixes.

+++ internal/infrastructure/networkpolicies.go
@@ -91,7 +91,7 @@ func dataplaneListenerNetworkPolicyPorts(
 			}
 
 			key := networkPolicyPortKey{
-				port:     int32(listener.Port),
+				port:     listener.Port,
 				protocol: protocol,
 			}
 			if _, exists := index[key]; exists {
